feat: combined all plugins
