+++ src/plugins/client/types/index.ts
@@ -1,0 +1,5 @@
+import { IIssuerOptions } from "../../../lib/types";
+
+export interface IClientPluginsOptions {
+    issuerOptions: IIssuerOptions
+}
