--- src/plugins/types/index.ts
@@ -1,9 +1,0 @@
-import { IIssuerOptions } from "../../lib/types";
-
-export interface ICorePluginsOptions {
-    issuerOptions: IIssuerOptions
-}
-
-export interface IClientPluginsOptions {
-    issuerOptions: IIssuerOptions
-}
