+++ src/plugins/vplay/index.ts
